Update patch set 2
Patch Set 2:

Building right now & I am going to test to see if they are. The issue with the old configuration is that it was unreliable at times. This is what is being used by Lineage and in OOS 3.1.4 so I am hoping to improve reliability... But then again, it might just be the poor renown connection we have here. I doubt it in this case though

Patch-set: 2
Label: Code-Review=0