onyx: prop: enable all cores
It has been proven that hotplugging cores is actually
a regression in battery life and performance. The case
remains the same here. The idea behind this is to only
utilize the mpdecision binary for all functions outside
of hotplugging so as to avoid a performance regression

This will also greatly aid with the upcoming fix for
frequency scaling

Change-Id: I6a1b8a797eadd87ac1741535b8d3df45e1db09cb