onyx: init: Slightly adjust input boost
Whenever the second core comes online, we should
make the assumption that is going to have to boost
to a higher frequency to deal with the given load

Also, 80ms seems a little too short for the boost
duration. 150-200ms seems optimal for boost periods
however with the 8974, each frequency step consumes
noticeably more power leading to greater heat in the
long run that isn't worth it, so let's try a slight
bump to 100ms instead.

Change-Id: I611bda15d9503f159a8b70d8c82690689379bfb6