Update patch set 1
Patch Set 1:

Sultans version of this driver seems to work slightly better in terms of UX while scrolling and stuff like that. I might investigate it later. For now, this is okay with me, but in the long run, I'd like to use the better solution. I'm not sure if this is any better with the genericized Qualcomm power HAL, but I'm not seeing it. I'll do proper comparisons, but for the time being, let's let it be

Patch-set: 1