Update patch set 1
Patch Set 1:

If the overlay isn't there because I don't have access to the closed stuff then its impossible for it to be added. I ported the kernel bit, and it works. But it is not possible for me to add pocket judge if I can't even use the overlay. I would be more than happy to otherwise. This is complete. Just not compilable

Patch-set: 1